added unit tests for all processes

+++ ERM.CSV.Extractor/Classes/LPFile.cs
@@ -6,14 +6,14 @@ using System.Threading.Tasks;
 
 namespace ERM.CSV.Extractor.Classes
 {
-    class LPFile
+     class LPFile
     {
         public double MeterPointCode { get; set; }
         public double SerialNumber { get; set; }
         public string PlantCode { get; set; }
         public DateTime DateAndTime { get; set; }
         public string DataType { get; set; }
-        public Decimal DataValue { get; set; }
+        public double DataValue { get; set; }
         public string Units { get; set; }
         public string Status { get; set; }
